darvis-mcp-server/pyproject.toml
[project]
name = "darvis-mcp-server"
version = "0.1.0"
description = "MCP tool server for DARVIS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "tavily-python>=0.5.0",
    "psutil>=6.0.0",
    "python-dateutil>=2.9.0",
    "dotenv>=0.9.9",
]
